updated the previous code with indicators.
Previous code will not auto connect to wifi if it disconnects. Now it will attempt to auto connect when wifi is disconnected. LED will blink every second when this happens. When its connected and functioning correctly, the LED will blink 0.5s every 5 seconds to indicate that the arduino is powered and ready.